ARM: dts: msm: bind rscc component to drm master for mtp and mtp-wcd9395
While transitioning from Primary VM to Trusted VM the number of
resources has to be same. RSCC resource is not bound to Primary
VM as its not present in connectors list of Primary vm device
tree. This change adds sde_rscc node in connectors list of
primary vm device tree for mtp and mtp-wcd935 target.

Change-Id: I3a4e7e5b03090780380714bde9dcdbe6f43aee0d